Comment out "before starting" steps in the cleaning instructions.
Due to the way protocols.io includes subprotocols, these steps would not have been seen by the OMI clinical team.
Since this AsciiDoc document is intended, for now, to produce something functionally identical to what the team is already using, the steps need to be removed.

Later on, we can add them back in.